refactor topic moderate
