+++ spirit/urls/__init__.py
@@ -10,6 +10,7 @@ urls = patterns('',
                 url(r'^st/admin/', include('spirit.urls.admin')),
                 url(r'^category/', include('spirit.urls.category')),
                 url(r'^topic/', include('spirit.urls.topic')),
+                url(r'^topic/moderate/', include('spirit.urls.topic_moderate')),
                 url(r'^topic/unread/', include('spirit.urls.topic_unread')),
                 url(r'^topic/notification/', include('spirit.urls.topic_notification')),
                 url(r'^topic/favorite/', include('spirit.urls.topic_favorite')),
