+++ spirit/urls/topic.py
@@ -15,31 +15,4 @@ urlpatterns = patterns("spirit.views.topic",
                        url(r'^(?P<pk>\d+)/(?P<slug>[\w-]+)/$', 'topic_detail', name='topic-detail'),
 
                        url(r'^active/$', 'topics_active', name='topic-active'),
-
-                       url(r'^delete/(?P<pk>\d+)/$',
-                           'topic_moderate',
-                           kwargs={'remove': True, 'value': True},
-                           name='topic-delete'),
-                       url(r'^undelete/(?P<pk>\d+)/$',
-                           'topic_moderate',
-                           kwargs={'remove': True, 'value': False},
-                           name='topic-undelete'),
-
-                       url(r'^lock/(?P<pk>\d+)/$',
-                           'topic_moderate',
-                           kwargs={'lock': True, 'value': True},
-                           name='topic-lock'),
-                       url(r'^unlock/(?P<pk>\d+)/$',
-                           'topic_moderate',
-                           kwargs={'lock': True, 'value': False},
-                           name='topic-unlock'),
-
-                       url(r'^pin/(?P<pk>\d+)/$',
-                           'topic_moderate',
-                           kwargs={'pin': True, 'value': True},
-                           name='topic-pin'),
-                       url(r'^unpin/(?P<pk>\d+)/$',
-                           'topic_moderate',
-                           kwargs={'pin': True, 'value': False},
-                           name='topic-unpin'),
                        )
